Refactor: Extract MSColab dialogs to separate modules
